+++ Viewer/resource.h
@@ -24,13 +24,16 @@
 #define ID_TOGGLE_DRAWER                32774
 #define ID_UPDATE                       32775
 #define ID_AUTOPLAY_VIDEOS              32776
+#define ID_SCALING_AUTO                 32777
+#define ID_SCALING_SMOOTH               32778
+#define ID_SCALING_PIXEL_EXACT          32779
 
 // Next default values for new objects
 //
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        140
-#define _APS_NEXT_COMMAND_VALUE         32777
+#define _APS_NEXT_COMMAND_VALUE         32780
 #define _APS_NEXT_CONTROL_VALUE         1004
 #define _APS_NEXT_SYMED_VALUE           101
 #endif